requirements.txt
faiss-cpu
sentence-transformers
openai
fastapi
uvicorn
httpx
requests
langchain
tiktoken
pydantic
transformers
pyperclip
watchdog
cookiecutter
typer
pandas
datasets
transformers==4.36.2
python-dotenv>=0.5.1
keyboard
pyautogui
pygetwindow